+++ src/scheduler-rules.js
@@ -4,7 +4,7 @@ const schedule = require('node-schedule');
 // Dreidev working days 10 am rule
 const workingDaysMoriningRule = new schedule.RecurrenceRule();
 workingDaysMoriningRule.dayOfWeek = [new schedule.Range(0, 4)];
-workingDaysMoriningRule.hour = 16;
-workingDaysMoriningRule.minute = 50;
+workingDaysMoriningRule.hour = 21;
+workingDaysMoriningRule.minute = 30;
 
 module.exports.workingDaysMoriningRule = workingDaysMoriningRule;
